archive/legacy/deploy_startup_fix_1752631218/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
email-validator==2.1.0
icalendar==5.0.11
flask==3.0.0
requests==2.31.0
langchain==0.1.0
langchain-google-genai==2.1.8
langgraph==0.2.50
langsmith==0.1.147
google-generativeai==0.8.3
typing-extensions==4.8.0
